==> 06_Train_random_forest.ipynb <==
# Train model on game and user data

In this notebook, we load in user and game information, compute some similarity scores between games and the games that usrs own, and then finally train a model to predict if users will own certain games. 

from sqlalchemy import create_engine
from sqlalchemy_utils import database_exists, create_database
import psycopg2
import pickle
import requests
import random
import re
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score, recall_score, precision_score
from sklearn.metrics import confusion_matrix
from sklearn.linear_model import LogisticRegression
from sklearn import preprocessing
from sklearn import metrics
from scipy.stats import randint as sp_randint
from sklearn.model_selection import RandomizedSearchCV
import ast
from gensim import corpora, models, similarities
from nltk.tokenize import word_tokenize
import math
import operator
import itertools
from time import time
from sklearn import cross_validation

%matplotlib inline
import matplotlib.pyplot as plt

# Define SQL database info
db_name  = 'UserInfo'
username = 'username'
host     = 'localhost'
pwd      = 'password'
port     = '5432'
engine = create_engine('postgresql://{}:{}@{}:{}/{}'.format(username, pwd, host, port, db_name))
print(engine.url)

# connect to database:
con = None
con = psycopg2.connect(database = db_name, user = username, password = pwd, host = host)
cur = con.cursor() # get a cursor to our current connection

## Query user data from the database

numUsers = 10000 # Number of users to train the model
# make a query to initialize the allPlayers table
create_table_sql = """
SELECT * FROM allPlayers WHERE gameCount IS NOT NULL ORDER BY random() LIMIT {};
""".format(numUsers)
userFrame = pd.read_sql_query(create_table_sql,con)

# For now we have to remove users that have the same game information (earlier error in writing data to database)
userFrame = userFrame.drop_duplicates(subset='games',keep="last")
len(userFrame)

## Prepare a dataframe with user information

testFrame = pd.DataFrame()
testFrame['timecreated'] = userFrame.timecreated
testFrame['usercountry'] = userFrame.loccountrycode
testFrame['games']       = userFrame.games

# Initialise lists to fill dataframe
numFriends   = []
numGames     = []
playTimeList = []
gameIDs      = []
for index, row in userFrame.iterrows():
    playTime   = 0 # Default
    gamesOwned = []
    try:
        friendDict = ast.literal_eval(row['friends'])
        numFriends.append(len(friendDict))
    except:
        numFriends.append(0) # If code throws an error, it is because user has no games
    try:
        gameDict = ast.literal_eval(row['games'])
        numGames.append(len(gameDict))
        for g in gameDict:
            playTime += g['playtime_forever']/60 # divide by 60 to get time expressed in hours
            gamesOwned.append(g['appid'])
        gameIDs.append(gamesOwned)
    except:
        numGames.append(0) # If code throws an error, it is because the user has no games
        gameIDs.append(0)
    playTimeList.append(playTime)    
        

# Now we fill in the columns of the dataframe with the lists we just compted
testFrame['friends']    = numFriends
testFrame['playTime']   = playTimeList
testFrame['numGames']   = numGames
testFrame['gamesOwned'] = gameIDs
testFrame = testFrame.dropna(axis=0,how='any')

## Load in TF-IDF corpus and dictionary

This will be used to compute a similarity score between games to be tested, and the games that each user currently owns.

# Load in the pretrained tf-idf models,dictionary, and corpus for game tags and game descriptions
tag_dictionary = corpora.Dictionary.load('/home/iain/Documents/InsghtProject/gameTagDict.dict')
tag_corpus     = corpora.MmCorpus('/home/iain/Documents/InsghtProject/gameTagCorpus.mm')
description_dictionary = corpora.Dictionary.load('/home/iain/Documents/InsghtProject/gameDescriptionDict.dict')
description_corpus     = corpora.MmCorpus('/home/iain/Documents/InsghtProject/gameDescriptionCorpus.mm')

# Initialise a term-frequency inverse document frequency model based on the corpus for both tags and descriptions
tag_tfidf = models.TfidfModel(tag_corpus) 
description_tfidf =  models.TfidfModel(description_corpus)
# Initialise a similarity index using the stored index, corpus, and dictionary
tag_sims = similarities.Similarity('/home/iain/Documents/InsghtProject/',
                                   tag_tfidf[tag_corpus],num_features=len(tag_dictionary))
description_sims = similarities.Similarity('/home/iain/Documents/InsghtProject/',
                    description_tfidf[description_corpus],num_features=len(description_dictionary))

## Load in game information

# Now let's load in the game data
# make a query to initialize the allGames table
create_table_sql = """
SELECT * FROM allGames WHERE gamename IS NOT NULL;
"""
gameFrame = pd.read_sql_query(create_table_sql,con)
gameFrame = gameFrame.dropna(axis=0, how='any') # Remove any rows that have nans
len(gameFrame)

## Define functions to compute similarity of game descriptions and tags

allGames = gameFrame.gameid.tolist()
def compute_tag_similarity(testGame,userGameDict):
    # This function comutes the similarity of one game to the games a user already owns based on game tags
    testRow     = gameFrame[gameFrame['gameid']==testGame]
    tmp         = testRow.tags.tolist()
    currentTags = ast.literal_eval(tmp[0]) # Evaluate to get in list form
    query       = ' '.join(currentTags) # Join all together with spaces

    # Tokenize
    query_doc = [w.lower() for w in word_tokenize(query)]
    # Compare to dictionary we computed on all other games
    query_doc_bow = tag_dictionary.doc2bow(query_doc)
    # Vectorize word representation
    query_doc_tf_idf = tag_tfidf[query_doc_bow]
    # Compute the similarity of the game with current tags to al games in our database
    query_similarity = tag_sims[query_doc_tf_idf]

    # Control for if the user has only one game
    if type(userGameDict) == dict:
        userGameDict = list([userGameDict])
        
    gamesOwned  = []
    gamesPlayed = []
    for g in userGameDict:
        if g['appid'] in allGames:
            if g['appid'] != testGame: # If the user owns the testGame, we should omit this from the model
                gamesOwned.append(g['appid'])
                gamesPlayed.append(g['playtime_forever']+0.01) # Add small amount to avoid dividing by zero
    # normalize the playtime to the total amount of playtime
    gamesPlayed = np.array([x / sum(gamesPlayed) for x in gamesPlayed]) 
    
    # If the user owns the testGame, we should omit this from the model
    if testGame in gamesOwned:
        gamesOwned.remove(testGame)
    
    # We only want to compare to games that are in our original list
    testInd = []
    for game in gamesOwned:
        if game in allGames:
            testInd.append(allGames.index(game))
    user2game_similarity = query_similarity[testInd]
    user2game_weighted_similarity = user2game_similarity*gamesPlayed
    
    # If we just have empty arrays, then fill with 0
    if user2game_similarity.size < 1:
        user2game_similarity = 0
        user2game_weighted_similarity = 0
        
    return np.median(user2game_similarity)#, np.median(user2game_weighted_similarity)

def preprocess_descriptions(description):
    # Define a list of words that we will not include, since they do not convey meaning
    try:
        removelist = set('for a of the and to in on is or be as where it its at an - with by'.split())
        description = re.sub(r"\s+", " ", description) # remove line breaks and tabs
        description = re.sub(r"<.*?>", "", description) # remove <>
        description = description.replace(",","") # remove commas
        description = description.replace(".","") # remove periods
        # Split text into a list of words
        text = [word for word in description.lower().split() if word not in removelist]
        return text
    except:
        return []

def compute_description_similarity(testGame,userGameDict):
    # This function comutes the similarity of one game to the games a user already owns based on game description
    testRow     = gameFrame[gameFrame['gameid']==testGame]
    tmp         = testRow.detailedescription.tolist()
    query_doc   = preprocess_descriptions(tmp[0]) # preprocess data

    # Compare to dictionary we computed on all other games
    query_doc_bow = description_dictionary.doc2bow(query_doc)
    # Vectorize word representation
    query_doc_tf_idf = description_tfidf[query_doc_bow]
    # Compute the similarity of the game with current tags to al games in our database
    query_similarity = description_sims[query_doc_tf_idf]
    
    # Control for if the user has only one game
    if type(userGameDict) == dict:
        userGameDict = list([userGameDict])
    
    gamesOwned  = []
    gamesPlayed = []
    for g in userGameDict:
        if g['appid'] in allGames:
            if g['appid'] != testGame: # If the user owns the testGame, we should omit this from the model
                gamesOwned.append(g['appid'])
                gamesPlayed.append(g['playtime_forever']+0.01)
    # normalize the playtime to the total amount of playtime
    gamesPlayed = np.array([x / sum(gamesPlayed) for x in gamesPlayed]) 
    
    # We only want to compare to games that are in our original list
    testInd = []
    for game in gamesOwned:
        if game in allGames:
            testInd.append(allGames.index(game))
    user2game_similarity = query_similarity[testInd]
    user2game_weighted_similarity = user2game_similarity*gamesPlayed
    
    # If we just have empty arrays, then fill with 0
    if user2game_similarity.size < 1:
        user2game_similarity = 0
        user2game_weighted_similarity = 0
    
    return np.median(user2game_similarity)#, np.median(user2game_weighted_similarity)

## Bring user and game data together in one frame

 Things to achieve in the cell below:
 - fill in a pandas dataframe with user information that we already have
 - into the same dataframe insert game inforamation
 - compute similarity index for tags and description
 - make dummy variables
 - control for class imbalance by undersampling

testData = testFrame[:numUsers]
test     = pd.DataFrame() # initialise dataframe

for index,user in testData.iterrows():    
    # Define the users game dict that contains games owned info and hours played
    gd = ast.literal_eval(user.games)

    # This code detects which games the user owns, and generates a list of games they don't own
    # To deal with imbalance in classes, we generate a list of games that uers don't own that is 
    # the same length of the list of games that they do own. 
    gmOwn = []
    for g in gd:
        if g['appid'] in allGames:
            gmOwn.append(g['appid'])
    gamesNotOwned = [x for x in allGames if x not in gmOwn]
    gamesNotOwned = random.choices(gamesNotOwned,k=len(gmOwn))
    
    if gmOwn: # We only want to process the user that have at least one game

        keepColumns = ['timecreated','usercountry','friends','playTime','numGames']
        userGames = gameFrame.loc[gameFrame['gameid'].isin(gmOwn)]
        userGames.loc[:,'userHasGame'] = pd.Series(np.ones(len(userGames)), index=userGames.index)
        testGames = gameFrame.loc[gameFrame['gameid'].isin(gamesNotOwned)]
        testGames.loc[:,'userHasGame'] = pd.Series(np.zeros(len(testGames)),index=testGames.index)
        tmp = pd.concat([userGames,testGames],axis=0)
        frame = pd.concat([pd.DataFrame(testData.loc[index,keepColumns]).T]*len(gmOwn)*2)

        # Reindex both tmp and frame
        tmp   = tmp.reset_index()
        frame = frame.reset_index()
        userPlusGame = pd.concat([frame,tmp],axis=1)
        
       # For some strange reason I was still getting some NaN's. let's just make sure they are not included
        userPlusGame = userPlusGame.dropna(axis=0, how='any') # Remove any rows that have nans
        
        # Now we compute the td-idf similarity scores 
        userPlusGame['description_similarity'] = userPlusGame.apply(lambda row: compute_description_similarity(row['gameid'], gd), axis=1)
        userPlusGame['tag_similarity'] = userPlusGame.apply(lambda row: compute_tag_similarity(row['gameid'], gd), axis=1)
         
        # Now I can append this to the end of the dataframe
        test = test.append(userPlusGame)

        print(index,end='\r')
    # We will test for each game in our game list, so we need to duplicate the user information n*games number of times
    #test = test.append([testData.loc[[index]]]*len(userGames)*2)
test = test.reset_index()

# Process dataframe to be suitable for modeling

Our model can only process numbers, so we have to convert categorical variables (country, genre, etc) into so-called 'dummy variables' which essentially converts a list of categorical variables into a much large sparse matrix of 0's and 1's (where each column now represents each instance of a categorical variable). 

# columns to delete:
# gameid
# index
# type
# gamename

def preprocess_columns(df):
    # We also have to remove any weird characters from the dataframe column names
    df.columns = [col.replace(',', '') for col in df.columns]
    df.columns = [col.replace('[', '') for col in df.columns]
    df.columns = [col.replace(']', '') for col in df.columns]
    df.columns = [col.replace('<', '') for col in df.columns]
    return df

dropColumns = ['gameid','index','type','gamename','is_free','level_0','detailedescription']
tempFrame   = test.drop(dropColumns, axis=1)

# Generate dummy dataframes from our categorical variables    
dfCountry   = pd.get_dummies(tempFrame['usercountry'])
dfCountry   = preprocess_columns(dfCountry)
dfDeveloper = pd.get_dummies(tempFrame['developers'])
dfDeveloper = preprocess_columns(dfDeveloper)
dfGenre     = pd.get_dummies(tempFrame['genre'])
dfGenre     = preprocess_columns(dfGenre)
#dfTags      = pd.get_dummies(tempFrame['tags'])

# Concatinate the dummy variables with the original dataframe
df_new      = pd.concat([tempFrame, dfCountry, dfDeveloper, dfGenre], axis=1)
# Remove the original categorical columns from the dataframe, keeping only the dummy variable columns
df_new = df_new.drop(['usercountry','developers','genre','tags'],axis=1)

# We have to convert the data type to float so that the XGBoost model can handle it
df_new['timecreated'] = pd.to_numeric(df_new['timecreated'],downcast='float')
df_new['friends'] = pd.to_numeric(df_new['friends'],downcast='float')
df_new['playTime'] = pd.to_numeric(df_new['playTime'],downcast='float')
df_new['numGames'] = pd.to_numeric(df_new['numGames'],downcast='float')

## Save dataframes

df_new.to_csv('/home/trainingDataset')
dfCountry.to_csv('/home/dfCountry')
dfDeveloper.to_csv('/home/dfDeveloper')
dfGenre.to_csv('/home/dfGenre')

df_new = pd.read_csv('/home/trainingDataset',nrows=100000)
del df_new['Unnamed: 0']
dfDeveloper = pd.read_csv('/home/dfDeveloper',nrows=1)
del dfDeveloper['Unnamed: 0']
dfCountry = pd.read_csv('/home/dfCountry',nrows=1)
del dfCountry['Unnamed: 0']
dfGenre = pd.read_csv('/home/dfGenre',nrows=1)
del dfGenre['Unnamed: 0']

## Make a list of features

dfCols     = pd.read_csv('/home/trainingDataset',nrows=1)
del dfCols['Unnamed: 0']
# Make a list of feature headers
featureHeaders = list(dfCols.columns)
targetFeature = 'userHasGame'
featureHeaders.remove(targetFeature)
mkList = []
for x in featureHeaders:
    mkList.append(x)  

## Make train test split for cross validation of models

# Split dataset into train and test dataset
train_percentage = 0.8
train_x, test_x, train_y, test_y = train_test_split(df_new[mkList], df_new[targetFeature],
                                                        train_size=train_percentage)

## Cross validate random forest model on data

clf = RandomForestClassifier(n_estimators=100,verbose=0)
scores = cross_validation.cross_val_score(clf, train_x, train_y, scoring = 'roc_auc',cv=10)
print("Print the scores for each fold of Random Forest cross validation:")
print(scores)

## Cross validate logistic regression model

logistic = LogisticRegression()
scores = cross_validation.cross_val_score(logistic, train_x, train_y, scoring = 'roc_auc',cv=10)
print("Print the scores for each fold of Logistic Regression cross validation:")
print(scores)

## Random forest model performs better!

As you can see above, the random forest model performs better than the logistic regression model. So we will push on and use a random forest model for our classification problem.

## Fit an ensemble of random forest models on segmented data

To avoid overfitting, we are going to fit an ensemble of models on different segments of our data. As you can see below, I am loading in the user/game dataframe in chunks of 20,000 rows. I then fit a random forest model on each set of 20k rows, append the model to a list, and then perform the same procedure for the next 20k rows. We perform this procedure to avoid overfitting a single model (or forest) on a particular subset of our data. 

chunkCount = -1
nchunks    = 10 # This is also the number of models in the ensemble
chunkSize  = 20000
modelList  = []

for i in range(nchunks):
    chunkCount = chunkCount + 1
    df_new = pd.read_csv('/home/trainingDataset',
        skiprows=range(1,chunkCount*chunkSize+1),nrows=chunkSize)
    del df_new['Unnamed: 0']
    df_new.columns = dfCols.columns
    
    # Initialise random forest model
    clf = RandomForestClassifier(n_estimators=100,verbose=3)
    clf.fit(df_new[mkList], df_new[targetFeature])
    modelList.append(clf)
    

## Define function to compute feature importance

Now that we are reasonably satisfied with the performance of the model, we can start to peek under the hood by looking at feature importances. In the code below, I define a function to compute the feature importance of our 'dummy variables' by summing the importances across all of the columns that make up each categorica variable. 

def computeImportance(headers,featDict):

    imp = 0
    for header in headers:
        #print(header)
        imp += featDict[header]
    return imp

# Collect feature importance for random forest model
rf_importances = np.zeros((len(modelList),len(mkList)))
for ind,clf in enumerate(modelList):
    rf_importances[ind,:] = clf.feature_importances_
rf_importance = np.mean(rf_importances,axis=0)

d = dict(zip(mkList,rf_importance))
countryHeaders           = list(dfCountry.columns)
d['countryImportance']   = computeImportance(countryHeaders,d)
genreHeaders             = list(dfGenre.columns)
d['genreImportance']     = computeImportance(genreHeaders,d)
developerHeaders         = list(dfDeveloper.columns)
d['developerImportance'] = computeImportance(developerHeaders,d)

## Plot feature importances

numFeatures = 12

sortedImportance = sorted(d.items(), key=operator.itemgetter(1),reverse=True)
top10 = sortedImportance[:numFeatures]

varName = []
varVal = []
for x in top10:
    varName.append(x[0])
    varVal.append(x[1])
xpos = np.flipud(range(len(top10)))

features=mkList
indices = np.argsort(rf_importance)

yticks = ['Developer studio','# Recommendations','User country','Release date','Description similarity',
         '# games user owns','Metacritic score','Playing time','Tag similarity','Time account created',
         '# friends','Game genre']

plt.figure(figsize=(7,5),dpi=100)
plt.title('Feature Importance',fontsize=20)
plt.barh(xpos, varVal, color='b', align='center',alpha=0.5)
plt.yticks(xpos, yticks,fontsize=20)
plt.xticks(fontsize=15)
plt.xlabel('Relative Importance',fontsize=20)
plt.grid()
plt.show()

## Save the ensemble of models

filename = '/home/ensemble_model.sav'
pickle.dump(modelList, open(filename, 'wb'))
